refactor: organiza bottom navigation bar
