+++ lib/screens/home_screen.dart
@@ -1,10 +1,9 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat/pages/calls_page.dart';
 import 'package:flutter_chat/pages/contacts_page.dart';
 import 'package:flutter_chat/pages/messages_page.dart';
 import 'package:flutter_chat/pages/notifications_page.dart';
-import 'package:flutter_chat/theme.dart';
+import 'package:flutter_chat/screens/widgets/bottom_navigation_bar.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -32,118 +31,8 @@ class _HomeScreenState extends State<HomeScreen> {
           return pages[value];
         },
       ),
-      bottomNavigationBar: _BottomNavigationBar(
+      bottomNavigationBar: BottomNavigationBarWidget(
         onItemSelected: (index) => pageIndex.value = index,
-      ),
-    );
-  }
-}
-
-class _BottomNavigationBar extends StatefulWidget {
-  final ValueChanged<int> onItemSelected;
-  const _BottomNavigationBar({required this.onItemSelected});
-
-  @override
-  State<_BottomNavigationBar> createState() => _BottomNavigationBarState();
-}
-
-class _BottomNavigationBarState extends State<_BottomNavigationBar> {
-  var selectedIndex = 0;
-
-  void handleItemSelected(int index) {
-    setState(() {
-      selectedIndex = index;
-    });
-    widget.onItemSelected;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-      top: false,
-      bottom: true,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          _NavigationBarItem(
-            index: 0,
-            label: 'Messages',
-            icon: CupertinoIcons.bubble_left_bubble_right_fill,
-            isSelected: (selectedIndex == 0),
-            onTap: handleItemSelected,
-          ),
-          _NavigationBarItem(
-            index: 1,
-            label: 'Notifications',
-            icon: CupertinoIcons.bell_solid,
-            isSelected: (selectedIndex == 1),
-            onTap: handleItemSelected,
-          ),
-          _NavigationBarItem(
-            index: 2,
-            label: 'Calls',
-            icon: CupertinoIcons.phone_fill,
-            isSelected: (selectedIndex == 2),
-            onTap: handleItemSelected,
-          ),
-          _NavigationBarItem(
-            index: 3,
-            label: 'Contacts',
-            icon: CupertinoIcons.person_fill,
-            isSelected: (selectedIndex == 3),
-            onTap: handleItemSelected,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _NavigationBarItem extends StatelessWidget {
-  final int index;
-  final String label;
-  final IconData icon;
-  final bool isSelected;
-  final ValueChanged<int> onTap;
-
-  const _NavigationBarItem({
-    required this.index,
-    required this.label,
-    required this.icon,
-    this.isSelected = false,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        onTap(index);
-      },
-      behavior: HitTestBehavior.opaque,
-      child: SizedBox(
-        height: 70,
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(
-              icon,
-              size: 20,
-              color: isSelected ? AppColors.secondary : null,
-            ),
-            const SizedBox(height: 8),
-            Text(
-              label,
-              style: isSelected
-                  ? const TextStyle(
-                      color: AppColors.secondary,
-                      fontWeight: FontWeight.bold,
-                      fontSize: 11,
-                    )
-                  : const TextStyle(fontSize: 11),
-            ),
-          ],
-        ),
       ),
     );
   }
