+++ lib/screens/widgets/bottom_navigation_bar.dart
@@ -1,0 +1,113 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter_chat/theme.dart';
+
+class BottomNavigationBarWidget extends StatefulWidget {
+  final ValueChanged<int> onItemSelected;
+  const BottomNavigationBarWidget({super.key, required this.onItemSelected});
+
+  @override
+  State<BottomNavigationBarWidget> createState() =>
+      BottomNavigationBarWidgetState();
+}
+
+class BottomNavigationBarWidgetState extends State<BottomNavigationBarWidget> {
+  var selectedIndex = 0;
+
+  void handleItemSelected(int index) {
+    setState(() {
+      selectedIndex = index;
+    });
+    widget.onItemSelected;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SafeArea(
+      top: false,
+      bottom: true,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceAround,
+        children: [
+          _NavigationBarItem(
+            index: 0,
+            label: 'Messages',
+            icon: CupertinoIcons.bubble_left_bubble_right_fill,
+            isSelected: (selectedIndex == 0),
+            onTap: handleItemSelected,
+          ),
+          _NavigationBarItem(
+            index: 1,
+            label: 'Notifications',
+            icon: CupertinoIcons.bell_solid,
+            isSelected: (selectedIndex == 1),
+            onTap: handleItemSelected,
+          ),
+          _NavigationBarItem(
+            index: 2,
+            label: 'Calls',
+            icon: CupertinoIcons.phone_fill,
+            isSelected: (selectedIndex == 2),
+            onTap: handleItemSelected,
+          ),
+          _NavigationBarItem(
+            index: 3,
+            label: 'Contacts',
+            icon: CupertinoIcons.person_fill,
+            isSelected: (selectedIndex == 3),
+            onTap: handleItemSelected,
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class _NavigationBarItem extends StatelessWidget {
+  final int index;
+  final String label;
+  final IconData icon;
+  final bool isSelected;
+  final ValueChanged<int> onTap;
+
+  const _NavigationBarItem({
+    required this.index,
+    required this.label,
+    required this.icon,
+    this.isSelected = false,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {
+        onTap(index);
+      },
+      behavior: HitTestBehavior.opaque,
+      child: SizedBox(
+        height: 70,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Icon(
+              icon,
+              size: 20,
+              color: isSelected ? AppColors.secondary : null,
+            ),
+            const SizedBox(height: 8),
+            Text(
+              label,
+              style: isSelected
+                  ? const TextStyle(
+                      color: AppColors.secondary,
+                      fontWeight: FontWeight.bold,
+                      fontSize: 11,
+                    )
+                  : const TextStyle(fontSize: 11),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
